+++ __tests__/node/utils/removeNonCodeWrappedHTML.spec.ts
@@ -1,3 +1,4 @@
+import { test, expect } from 'vitest'
 import { removeNonCodeWrappedHTML } from 'node/utils/parseHeader'
 
 test('removeNonCodeWrappedHTML', () => {