+++ src/client/shim.d.ts
@@ -1,7 +1,7 @@
 declare const __VP_HASH_MAP__: Record<string, string>
-declare const __CARBON__: boolean
-declare const __BSA__: boolean
 declare const __ALGOLIA__: boolean
+declare const __CARBON__: boolean
+
 declare module '*.vue' {
   import { ComponentOptions } from 'vue'
   const comp: ComponentOptions
@@ -13,10 +13,14 @@ declare module '@siteData' {
   export default data
 }
 
-// this module's typing is broken
+// this module's typing is broken.
 declare module '@docsearch/js' {
   function docsearch<T = any>(props: T): void
   export default docsearch
+}
+
+declare module '@docsearch/react/dist/esm/types' {
+  export type DocSearchHit = any
 }
 
 declare module '@docsearch/css' {