+++ src/client/theme-default/index.ts
@@ -1,14 +1,23 @@
+import './styles/fonts.css'
 import './styles/vars.css'
-import './styles/layout.css'
-import './styles/code.css'
-import './styles/custom-blocks.css'
-import './styles/sidebar-links.css'
+import './styles/base.css'
+import './styles/utils.css'
+import './styles/components/custom-block.css'
+import './styles/components/vp-code.css'
+import './styles/components/vp-doc.css'
+import './styles/components/vp-sponsor.css'
 
 import { Theme } from 'vitepress'
 import Layout from './Layout.vue'
 import NotFound from './NotFound.vue'
 
 export { DefaultTheme } from './config'
+
+export { default as VPHomeHero } from './components/VPHomeHero.vue'
+export { default as VPHomeFeatures } from './components/VPHomeFeatures.vue'
+export { default as VPHomeSponsors } from './components/VPHomeSponsors.vue'
+export { default as VPDocAsideSponsors } from './components/VPDocAsideSponsors.vue'
+
 const theme: Theme = {
   Layout,
   NotFound