+++ src/node/markdown/plugins/highlight.ts
@@ -1,52 +1,28 @@
-import chalk from 'chalk'
-import escapeHtml from 'escape-html'
-import prism from 'prismjs'
+import { getHighlighter } from 'shiki'
+import type { ThemeOptions } from '../markdown'
 
-// prism is listed as actual dep so it's ok to require
-const loadLanguages = require('prismjs/components/index')
+export async function highlight(theme: ThemeOptions = 'material-palenight') {
+  const themes = typeof theme === 'string' ? [theme] : [theme.dark, theme.light]
+  const highlighter = await getHighlighter({ themes })
+  const preRE = /^<pre.*?>/
 
-// required to make embedded highlighting work...
-loadLanguages(['markup', 'css', 'javascript'])
+  return (str: string, lang: string) => {
+    lang = lang || 'text'
 
-function wrap(code: string, lang: string): string {
-  if (lang === 'text') {
-    code = escapeHtml(code)
-  }
-  return `<pre v-pre><code>${code}</code></pre>`
-}
-
-export const highlight = (str: string, lang: string) => {
-  if (!lang) {
-    return wrap(str, 'text')
-  }
-  lang = lang.toLowerCase()
-  const rawLang = lang
-  if (lang === 'vue' || lang === 'html') {
-    lang = 'markup'
-  }
-  if (lang === 'md') {
-    lang = 'markdown'
-  }
-  if (lang === 'ts') {
-    lang = 'typescript'
-  }
-  if (lang === 'py') {
-    lang = 'python'
-  }
-  if (!prism.languages[lang]) {
-    try {
-      loadLanguages([lang])
-    } catch (e) {
-      console.warn(
-        chalk.yellow(
-          `[vitepress] Syntax highlight for language "${lang}" is not supported.`
-        )
-      )
+    if (typeof theme === 'string') {
+      return highlighter
+        .codeToHtml(str, { lang, theme })
+        .replace(preRE, '<pre v-pre>')
     }
+
+    const dark = highlighter
+      .codeToHtml(str, { lang, theme: theme.dark })
+      .replace(preRE, '<pre v-pre class="vp-code-dark">')
+
+    const light = highlighter
+      .codeToHtml(str, { lang, theme: theme.light })
+      .replace(preRE, '<pre v-pre class="vp-code-light">')
+
+    return dark + light
   }
-  if (prism.languages[lang]) {
-    const code = prism.highlight(str, prism.languages[lang], lang)
-    return wrap(code, rawLang)
-  }
-  return wrap(str, 'text')
 }