+++ src/shared/shared.ts
@@ -1,4 +1,4 @@
-import { LocaleConfig, SiteData } from '../../types/shared'
+import { SiteData, PageData, LocaleConfig } from '../../types/shared'
 
 export type {
   SiteData,
@@ -10,6 +10,7 @@ export type {
 } from '../../types/shared'
 
 export const EXTERNAL_URL_RE = /^https?:/i
+export const APPEARANCE_KEY = 'vitepress-theme-appearance'
 
 // @ts-ignore
 export const inBrowser = typeof window !== 'undefined'
@@ -80,6 +81,36 @@ export function resolveSiteDataByRoute(
     locales: {},
     langs: createLangDictionary(siteData)
   })
+}
+
+/**
+ * Create the page title string based on configs.
+ */
+export function createTitle(siteData: SiteData, pageData: PageData): string {
+  const title = pageData.title || siteData.title
+  const template = pageData.titleTemplate ?? siteData.titleTemplate
+  const templateString = createTitleTemplate(siteData.title, template)
+
+  return `${title}${templateString}`
+}
+
+function createTitleTemplate(
+  siteTitle: string,
+  template?: string | boolean
+): string {
+  if (template === false) {
+    return ''
+  }
+
+  if (template === true || template === undefined) {
+    return ` | ${siteTitle}`
+  }
+
+  if (siteTitle === template) {
+    return ''
+  }
+
+  return ` | ${template}`
 }
 
 /**