+++ types/shared.d.ts
@@ -5,25 +5,38 @@ export { DefaultTheme } from './default-theme'
 export interface PageData {
   relativePath: string
   title: string
+  titleTemplate?: string | boolean
   description: string
   headers: Header[]
   frontmatter: Record<string, any>
   lastUpdated?: number
 }
 
+export interface Header {
+  level: number
+  title: string
+  slug: string
+}
+
 export interface SiteData<ThemeConfig = any> {
   base: string
+
   /**
    * Language of the site as it should be set on the `html` element.
+   *
    * @example `en-US`, `zh-CN`
    */
   lang: string
+
   title: string
+  titleTemplate?: string | boolean
   description: string
   head: HeadConfig[]
+  appearance: boolean
   themeConfig: ThemeConfig
   scrollOffset: number | string
   locales: Record<string, LocaleConfig>
+
   /**
    * Available locales for the site when it has defined `locales` in its
    * `themeConfig`. This object is otherwise empty. Keys are paths like `/` or
@@ -50,15 +63,10 @@ export type HeadConfig =
   | [string, Record<string, string>]
   | [string, Record<string, string>, string]
 
-export interface Header {
-  level: number
-  title: string
-  slug: string
-}
-
 export interface LocaleConfig {
   lang: string
   title?: string
+  titleTemplate?: string | boolean
   description?: string
   head?: HeadConfig[]
   label?: string